fix: resolve backend-size limits and remove dead Codex adapter helper
